--- frontend/src/types/api/ShiftSignupStatus.ts
@@ -1,5 +1,0 @@
-export type ShiftSignupStatus =
-  | "PENDING"
-  | "CONFIRMED"
-  | "CANCELED"
-  | "PUBLISHED";
